projects/watchdog/lambda_scheduled/requirements.txt
uuid1
simplejson
boto3
python-dateutil